+++ app/geometry.ts
@@ -1,6 +1,32 @@
+import { Position } from 'geojson';
 import RBush, { BBox } from 'rbush';
 import Queue from 'tinyqueue';
-import { compareDistance } from './tree';
+
+export function compareDistance(
+  a: { distance: number },
+  b: { distance: number }
+) {
+  return a.distance - b.distance;
+}
+
+export function pointDistance(a: Position, b: Position) {
+  const dx = a[0] - b[0];
+  const dy = a[1] - b[1];
+  return dx * dx + dy * dy;
+}
+
+export function pointLineSegmentDistance(
+  c: Position,
+  a: Position,
+  b: Position,
+  d: typeof pointDistance = pointDistance
+) {
+  const dx = b[0] - a[0];
+  const dy = b[1] - a[1];
+  const d2 = dx * dx + dy * dy;
+  const t = d2 && ((c[0] - a[0]) * dx + (c[1] - a[1]) * (b[1] - a[1])) / d2;
+  return d(c, t <= 0 ? a : t >= 1 ? b : [a[0] + t * dx, a[1] + t * dy]);
+}
 
 // function boxDist(x: number, y: number, box: BBox) {
 //   const dx = axisDist(x, box.minX, box.maxX),
