+++ app/tree.ts
@@ -4,203 +4,3 @@
 // FIXME license?
 
 import { Position } from 'geojson';
-import minHeap from './min-heap';
-
-export function compareDistance(
-  a: { distance: number },
-  b: { distance: number }
-) {
-  return a.distance - b.distance;
-}
-
-export function pointDistance(a: Position, b: Position) {
-  const dx = a[0] - b[0];
-  const dy = a[1] - b[1];
-  return dx * dx + dy * dy;
-}
-
-export function pointLineSegmentDistance(
-  c: Position,
-  a: Position,
-  b: Position,
-  d: typeof pointDistance = pointDistance
-) {
-  const dx = b[0] - a[0];
-  const dy = b[1] - a[1];
-  const d2 = dx * dx + dy * dy;
-  const t = d2 && ((c[0] - a[0]) * dx + (c[1] - a[1]) * (b[1] - a[1])) / d2;
-  return d(c, t <= 0 ? a : t >= 1 ? b : [a[0] + t * dx, a[1] + t * dy]);
-}
-
-export function group<T>(children: TreeNode<T>[]): Node<T> {
-  let n0;
-
-  while ((n0 = children.length) > 1) {
-    const children1 = new Array(Math.ceil(n0 / 2));
-    let i0, i1;
-    for (i0 = 0, i1 = 0; i0 < n0 - 1; i0 += 2, i1 += 1) {
-      const child0 = children[i0];
-      const child1 = children[i0 + 1];
-      children1[i1] = new Node(child0, child1);
-    }
-
-    if (i0 < n0) {
-      children1[i1] = children[i0];
-    }
-
-    children = children1;
-  }
-
-  return children[0] as any;
-}
-
-export type TreeNode<T> = Node<T> | Leaf<T>;
-
-type HeapEntry<T> = { node: TreeNode<T>; distance: number };
-
-export class Node<T> {
-  extent: [Position, Position];
-  children: [TreeNode<T>, TreeNode<T>];
-  constructor(child0: TreeNode<T>, child1: TreeNode<T>) {
-    const e0 = child0.extent;
-    const e1 = child1.extent;
-    this.children = [child0, child1];
-    this.extent = [
-      [Math.min(e0[0][0], e1[0][0]), Math.min(e0[0][1], e1[0][1])],
-      [Math.max(e0[1][0], e1[1][0]), Math.max(e0[1][1], e1[1][1])],
-    ];
-  }
-
-  distance(point: Position, d: typeof pointDistance): number {
-    const [x, y] = point;
-    const [[x0, y0], [x1, y1]] = this.extent;
-
-    return x < x0
-      ? pointLineSegmentDistance(point, [x0, y0], [x0, y1], d)
-      : x > x1
-      ? pointLineSegmentDistance(point, [x1, y0], [x1, y1], d)
-      : y < y0
-      ? pointLineSegmentDistance(point, [x0, y0], [x1, y0], d)
-      : y > y1
-      ? pointLineSegmentDistance(point, [x0, y1], [x1, y1], d)
-      : 0;
-  }
-
-  nearestWithDistance(
-    point: Position,
-    d: typeof pointDistance = pointDistance,
-    closeEnough = -Infinity,
-    tooFar = Infinity
-  ): { node: Leaf<T>; distance: number } {
-    let minNode: Leaf<T>;
-    let minDistance = Infinity;
-    const heap = minHeap<HeapEntry<T>>(compareDistance);
-    let node: TreeNode<T> = this;
-    let candidate: HeapEntry<T> | null = {
-      distance: node.distance(point, d),
-      node,
-    };
-
-    do {
-      node = candidate.node;
-      if (node.children) {
-        let distance = node.children[0].distance(point, d);
-        if (distance < tooFar) {
-          heap.push({ node: node.children[0], distance });
-        }
-        distance = node.children[1].distance(point, d);
-        if (distance < tooFar) {
-          heap.push({ node: node.children[1], distance });
-        }
-      } else {
-        const distance = node.distance(point, d);
-        if (distance <= closeEnough) {
-          return { node, distance };
-        }
-        if (distance < minDistance) {
-          minDistance = distance;
-          minNode = node;
-        }
-      }
-    } while ((candidate = heap.pop()) && candidate.distance <= minDistance);
-
-    return { node: minNode!, distance: minDistance };
-  }
-
-  nearest(point: Position, d: typeof pointDistance = pointDistance): Leaf<T> {
-    return this.nearestWithDistance(point, d).node;
-  }
-
-  within(
-    point: Position,
-    maxDistance: number,
-    d: typeof pointDistance = pointDistance
-  ): { node: Leaf<T>; distance: number }[] {
-    const result: { node: Leaf<T>; distance: number }[] = [];
-    const visit = (node: TreeNode<T>) => {
-      const distance = node.distance(point, d);
-      if (distance <= maxDistance) {
-        if (!isLeaf(node)) {
-          node.children.map(visit);
-        } else {
-          result.push({ node, distance });
-        }
-      }
-    };
-
-    visit(this);
-
-    return result;
-  }
-}
-
-function isLeaf<T>(n: TreeNode<T>): n is Leaf<T> {
-  return !n.children;
-}
-
-export class Leaf<T> {
-  coordinates: [Position, Position];
-  extent: [Position, Position];
-  children: undefined;
-  constructor(
-    point0: Position,
-    point1: Position,
-    public index: number,
-    public data: T
-  ) {
-    this.coordinates = [point0, point1];
-    this.extent = [
-      [Math.min(point0[0], point1[0]), Math.min(point0[1], point1[1])],
-      [Math.max(point0[0], point1[0]), Math.max(point0[1], point1[1])],
-    ];
-  }
-
-  distance(point: Position, d: typeof pointDistance): number {
-    return pointLineSegmentDistance(
-      point,
-      this.coordinates[0],
-      this.coordinates[1],
-      d
-    );
-  }
-}
-
-export default function <T>(arcs: { arc: Position[]; data: T }[]): Node<T> {
-  return group(
-    arcs.map(({ arc, data }) => {
-      let i = 0;
-      const n = arc.length;
-      let p0;
-      let p1 = arc[0];
-      const children: Leaf<T>[] = new Array(n - 1);
-
-      while (++i < n) {
-        p0 = p1;
-        p1 = arc[i];
-        children[i - 1] = new Leaf(p0, p1, i, data);
-      }
-
-      return group(children);
-    })
-  );
-}
